requirements.txt
streamlit
requests
pdfplumber
numpy
beautifulsoup4
sentence-transformers
google-generativeai
supabase
python-dotenv
